firehose.url_id now int, not mediumint
urls.url_id was already int

tracks internal commit affeae87903808dd36c6e52dd5f3aeecd2dd51f7

+++ plugins/FireHose/mysql_schema.sql
@@ -18,7 +18,7 @@ CREATE TABLE firehose (
 	primaryskid smallint(5) default '0',
 	tid smallint(6) default '0',
 	srcid int(8) unsigned NOT NULL default '0',
-	url_id mediumint(8) unsigned NOT NULL default '0',
+	url_id int(8) unsigned NOT NULL default '0',
 	toptags varchar(255) default '',
 	email varchar(255) NOT NULL default '',
 	emaildomain varchar(255) NOT NULL default '',
